move persistentdict into own file, rename
